+++ src/Yasumi/SubstituteHoliday.php
@@ -68,14 +68,14 @@ class SubstituteHoliday extends Holiday
     ) {
         $this->substitutedHoliday = $substitutedHoliday;
 
-        $shortName = 'substituteHoliday:' . $substitutedHoliday->getShortName();
+        $key = 'substituteHoliday:' . $substitutedHoliday->getKey();
 
         if ($date == $substitutedHoliday) {
             throw new \InvalidArgumentException('Date must differ from the substituted holiday');
         }
 
         // Construct instance
-        parent::__construct($shortName, $names, $date, $displayLocale, $type);
+        parent::__construct($key, $names, $date, $displayLocale, $type);
     }
 
     /**
@@ -94,20 +94,20 @@ class SubstituteHoliday extends Holiday
      * The provided locales are searched for a translation. The first locale containing a translation will be used.
      *
      * If no locale is provided, proceed as if an array containing the display locale, Holiday::DEFAULT_LOCALE ('en_US'), and
-     * Holiday::LOCALE_SHORT_NAME (the short name (internal name) of this holiday) was provided.
+     * Holiday::LOCALE_KEY (the holiday key) was provided.
      *
      * @param array $locales The locales to search for translations
      *
      * @throws MissingTranslationException
      *
      * @see Holiday::DEFAULT_LOCALE
-     * @see Holiday::LOCALE_SHORT_NAME
+     * @see Holiday::LOCALE_KEY
      */
     public function getName($locales = null): string
     {
         $name = parent::getName();
 
-        if ($name === $this->getShortName()) {
+        if ($name === $this->getKey()) {
             foreach ($this->getLocales($locales) as $locales) {
                 $pattern = $this->substituteHolidayTranslations[$locales] ?? null;
                 if ($pattern) {