+++ tests/Ukraine/SubstitutedHolidayTest.php
@@ -59,7 +59,7 @@ class SubstitutedHolidayTest extends UkraineBaseTestCase implements YasumiTestCa
      * Asserts that the expected date is indeed a holiday for that given year and name
      *
      * @param string $provider the holiday provider (i.e. country/state) for which the holiday need to be tested
-     * @param string $shortName string the short name of the holiday to be checked against
+     * @param string $key string the key of the holiday to be checked against
      * @param int $year holiday calendar year
      * @param DateTime $expectedOfficial the official date to be checked against
      * @param DateTime $expectedSubstitution the substituted date to be checked against
@@ -73,21 +73,21 @@ class SubstitutedHolidayTest extends UkraineBaseTestCase implements YasumiTestCa
      */
     public function assertHolidayWithSubstitution(
         string $provider,
-        string $shortName,
+        string $key,
         int $year,
         DateTime $expectedOfficial,
         DateTime $expectedSubstitution = null
     ): void {
         $holidays = Yasumi::create($provider, $year);
 
-        $holidayOfficial = $holidays->getHoliday($shortName);
+        $holidayOfficial = $holidays->getHoliday($key);
         $this->assertInstanceOf(Holiday::class, $holidayOfficial);
         $this->assertNotNull($holidayOfficial);
         $this->assertEquals($expectedOfficial, $holidayOfficial);
         $this->assertTrue($holidays->isHoliday($holidayOfficial));
         $this->assertEquals(Holiday::TYPE_OFFICIAL, $holidayOfficial->getType());
 
-        $holidaySubstitution = $holidays->getHoliday('substituteHoliday:' . $holidayOfficial->getShortName());
+        $holidaySubstitution = $holidays->getHoliday('substituteHoliday:' . $holidayOfficial->getKey());
         if ($expectedSubstitution === null) {
             // without substitution
             $this->assertNull($holidaySubstitution);